Move Paragraph property to own class.
